Update how we calculate time to merge for a dependabot PR
At the moment we are ignoring any superseded Dependabot PRs in time-to-merge calculation.
We want to have a more accurate metric on how long it takes until
a dependency is upgraded to the latest version.

The changes include:
- use list_issues for fetching dependabot PRs for each repo.
In this way we retrieve results only for govuk repos (it is limited at 30 PRs per repo,
but this can be changed). And we are no longer restricted by the Search API rate-limiting.
- the time since the dependency is outdated is calculated using the PR history of the repo.
For each dependency, we are getting the earliest date a pr was opened
for bumping the current version.
The limitation for this logic is that `list_issues` returns only 30 prs per page,
and we might miss some days for dependencies that are outdated for a very long time.